% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_OFX.R
\name{write_OFX}
\alias{write_OFX}
\title{Write OFX}
\usage{
write_OFX(statement, outfile, date, funds,
  header = "<?xml version=\\"1.0\\" encoding=\\"UTF-8\\"?>\\n<?OFX OFXHEADER=\\"200\\" VERSION=\\"200\\" SECURITY=\\"NONE\\" OLDFILEUID=\\"NONE\\" NEWFILEUID=\\"NONE\\"?>\\n")
}
\arguments{
\item{statement}{data.frame. Questrade statement formatted using the `format_statement` function}

\item{outfile}{character. Output file path. Should have '.ofx' extension.}

\item{date}{POSIXct. Statement date and time.}

\item{funds}{data.frame. A table containing the name and symbol of all the funds on the statement}

\item{header}{character. XML header. Default header used for Microsoft Money.}
}
\description{
Write formated Questrade statement to an OFX file. Input should come from the `format_statement` function.
}
